Ascension's 32 classes, and four latent bugs their data triggers
Everything the local Ascension realm needs from the core, on a branch instead of
as an uncommitted working-tree diff on upstream master, where one git checkout
would have destroyed it. The same diff is carried in
~/repos/ascender-server/patches/mangostwo-server.patch so the realm can be
rebuilt from that repository alone.

MAX_CLASSES and MAX_RACES were 12; Sun Cleric is class 27, so they are 33 and 64
and CLASSMASK_ALL_PLAYABLE is widened to all 32 bits, or playercreateinfo drops
the row with "Wrong class 27".

Three of these are latent MaNGOS bugs that stock data simply never reaches. A
third-party Spell.dbc contradicts its own rank chains, and 223 links here do, so
four MANGOS_ASSERTs became counted drops. A gap in TaxiPathNode.dbc leaves a
default TaxiPathNodePtr whose conversion operator dereferences its NULL
unconditionally -- one null among 23,314 rows, surfacing as a SIGSEGV in
ScriptMgr::CollectPossibleEventIds, a function with nothing to do with taxi
paths. An assert on achievement criteria types outside this build's enum also
indexed an array out of bounds, so those rows had to be dropped either way.

TalentSpellPos::talent_id was a uint16 while Talent.dbc's id column is 32 bits
and Ascension's runs to 112369: talent 112286 came back as 46750, a real id
belonging to a different talent, so every talent above 65535 was silently
mis-mapped. Widened, with null checks on the lookups that were dereferenced
unconditionally beside it.

Talent points for a class above MAX_STOCK_CLASS grow with level and are one pool
spendable anywhere, gated at TALENT_POINTS_PER_TIER_ADDED per tier. The fixed
per-tree pool and its per-tree cap that shipped before made tier 1 need five
points spent and refuse the sixth, so an added class could pick its first row
and nothing else, at any level, forever.

player_class_levelspell is new: playercreateinfo_spell is creation-only and
there are no trainers for these classes, so without it a class learned
everything it would ever know at creation.

And the session key is hashed at the protocol's 40 bytes rather than the value's
minimal encoding, in AuthCrypt and the CMSG_AUTH_SESSION digest. Those agree
only while the key's top byte is non-zero; once in 256 the server hashed 39
bytes against the client's 40 and refused the login as a bad proof. That one is
upstream's and worth a pull request.

+++ src/game/Server/DBCStructure.h
@@ -2619,12 +2619,24 @@ typedef std::set<uint32> PetFamilySpellsSet;
 typedef std::map<uint32, PetFamilySpellsSet > PetFamilySpellsStore;
 
 // Structures not used for casting to loaded DBC data and not required then packing
+/**
+ * Where a spell sits in the talent trees: which Talent.dbc row grants it, and
+ * at which rank.
+ *
+ * talent_id IS A uint32, not a uint16. Talent.dbc's id column is 32 bits and
+ * stock 3.3.5a happens to stay under 65535, so the narrower field was free --
+ * until a DBC arrives that does not. Ascension's Talent.dbc runs to id 112369,
+ * and 112286 stored in a uint16 comes back as 46750: a real id, belonging to a
+ * different talent, which the server then cannot find or finds as the wrong
+ * one. The failure is silent and looks like the talent simply refusing to be
+ * learned.
+ */
 struct TalentSpellPos
 {
     TalentSpellPos() : talent_id(0), rank(0) {}
-    TalentSpellPos(uint16 _talent_id, uint8 _rank) : talent_id(_talent_id), rank(_rank) {}
+    TalentSpellPos(uint32 _talent_id, uint8 _rank) : talent_id(_talent_id), rank(_rank) {}
 
-    uint16 talent_id;
+    uint32 talent_id;
     uint8  rank;
 };
